+++ packages/ace-core/package.json
@@ -26,7 +26,6 @@
     "axe-core": "^3.2.2",
     "file-url": "^2.0.2",
     "h5o": "^0.11.3",
-    "i18next": "^15.1.0",
     "p-map": "^1.2.0",
     "puppeteer": "^1.0.0",
     "tmp": "^0.0.33",
