+++ packages/ace-localize/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@daisy/ace-report",
-  "version": "1.0.2",
-  "description": "Reporting utilities for Ace",
+  "name": "@daisy/ace-localize",
+  "version": "1.0.0",
+  "description": "Localization for Ace",
   "author": {
     "name": "DAISY developers",
     "organization": "DAISY Consortium",
@@ -15,13 +15,9 @@
     "url": "https://github.com/daisy/ace/issues"
   },
   "license": "MIT",
-  "main": "lib/index.js",
+  "main": "lib/localize.js",
   "dependencies": {
-    "@daisy/ace-config": "^1.0.0",
-    "@daisy/ace-meta": "^1.0.3",
-    "escape-html": "^1.0.3",
-    "fs-extra": "^6.0.1",
-    "handlebars": "^4.0.11",
+    "i18next": "^15.1.0",
     "winston": "^2.4.0"
   },
   "publishConfig": {
