+++ packages/ace-logger/src/index.js
@@ -25,8 +25,8 @@ module.exports.initLogger = function initLogger(options = {}) {
   // TODO remove this, this is just for testing/debugging
   if (options.verbose && !options.silent) {
     console.log(`>>>>> LOG FILE PATH: ${logfile}`);
-    process.stdout.write(`>>>>> LOG FILE PATH: ${logfile}`);
-    process.stderr.write(`>>>>> LOG FILE PATH: ${logfile}`);
+    process.stdout.write(`>>>>> LOG FILE PATH: ${logfile}\n`);
+    process.stderr.write(`>>>>> LOG FILE PATH: ${logfile}\n`);
   }
 
   // set up logger
