+++ packages/ace-report-axe/package.json
@@ -19,7 +19,6 @@
   "dependencies": {
     "@daisy/ace-report": "^1.0.1",
     "fs-extra": "^6.0.1",
-    "i18next": "^15.1.0",
     "winston": "^2.4.0"
   },
   "publishConfig": {
