+++ packages/ace-report/src/report.js
@@ -8,7 +8,7 @@ const builders = require('./report-builders.js');
 const a11yMetaChecker = require("./analyze-a11y-metadata.js");
 const generateHtmlReport = require("./generate-html-report.js");
 
-const { localize, setCurrentLanguage } = require('./l10n/localize');
+const { localize, setCurrentLanguage } = require('./l10n/localize').localizer;
 
 function headingsToOutline(headings) {
   const result = [];
